# specs/001-shoktiai-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ShoktiAI Backend API
  version: 1.0.0
  description: Core APIs for Customer, Worker, Admin, and internal AI orchestration
servers:
  - url: https://api.example.com
paths:
  /auth/request-otp:
    post:
      summary: Request OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, example: "+8801XXXXXXXXX" }
      responses:
        '200': { description: OTP sent }
  /auth/verify-otp:
    post:
      summary: Verify OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string }
                code: { type: string }
      responses:
        '200':
          description: Auth token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
...